% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_import.R
\name{bw_import}
\alias{bw_import}
\title{Import BigWig region}
\usage{
bw_import(bw_file, chrom, start, end)
}
\arguments{
\item{bw_file}{Character scalar: path to a local BigWig or a URL (http/https/ftp)}

\item{chrom}{Character scalar: chromosome name (e.g., "chr1")}

\item{start}{Integer(1): 1-based start (inclusive)}

\item{end}{Integer(1): 1-based end (inclusive)}
}
\value{
Numeric vector of length end - start + 1
}
\description{
Import BigWig region
}
\examples{
bw_URL = "http://genome-ftp.mbg.au.dk/public/THJ/seqNdisplayR/examples/tracks/HeLa_3pseq/siGFP_noPAP_in_batch1_plus.bw"
chrom = "chr12"
chrom_start = 6531808
chrom_end = 6541078
vals = bw_import(bw_URL, chrom, chrom_start, chrom_end)
max(vals) # [1] 9503.29
}
