% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_import.R
\name{bw_import}
\alias{bw_import}
\title{Import BigWig region}
\usage{
bw_import(bw_file, chrom, start, end)
}
\arguments{
\item{bw_file}{path or URL}

\item{chrom}{chromosome (e.g., "chr12")}

\item{start}{1-based start}

\item{end}{1-based end inclusive}
}
\value{
numeric vector of length end-start+1
}
\description{
Import BigWig region
}
